src/skyllian/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
ecdsa==0.18.0
flake8==5.0.4
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
pycodestyle==2.9.1
pycryptodome==3.15.0
pyflakes==2.5.0
pymongo==4.3.3
requests==2.28.1
six==1.16.0
textdistance==4.5.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
pika==1.3.1
